+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The complete changelog for the Costs to Expect REST API, follows the format defi
 ### Added 
 - We have added `X-Sort` and `X-Search` headers to the `/v1/resource-types/[resource-type]/items` collection.
 - We have added `source` to the `/v1/request/error-log` collection.
+- We have added listeners to revoke and prune old access tokens.
 
 ### Fixed 
 - Sorting the `/v1/resource-type/[resource-type]/items` collection by description generates an error, table name missing from sort field causing ambiguity.